added pop-up for fitting
